requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
drf-yasg>=1.21.0
psycopg2-binary>=2.9.5
pytest>=7.0.0
pytest-django>=4.5.0
python-dotenv>=1.0.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt
djangorestframework
django-filter
pytest
pytest-django
psycopg2-binary
drf-yasg
django-redis